refactor: configure layout correctly
