+++ Client/streamverse-app/src/app/(auth)/register/page.tsx
@@ -6,10 +6,11 @@ import { Label } from "@/components/ui/label";
 import { Button } from "@/components/ui/button";
 import { AuthCard } from "@/components/auth/AuthCard";
 import { GoogleButton } from "@/components/auth/GoogleButton";
-import { handleLogin } from "@/lib/auth";
+import { handleRegister } from "@/lib/auth";
 import { Loader2 } from "lucide-react";
 
-export default function LoginPage() {
+export default function RegisterPage() {
+  const [username, setUsername] = useState("");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [isLoading, setIsLoading] = useState(false);
@@ -17,19 +18,31 @@ export default function LoginPage() {
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
     setIsLoading(true);
-    await handleLogin({ email, password });
+    await handleRegister({ username, email, password });
     setIsLoading(false);
   };
 
   return (
     <AuthCard
-      title="Login"
-      description="Welcome back! Please enter your details."
-      footerText="Don't have an account?"
-      footerLink="/auth/register"
-      footerLinkText="Register"
+      title="Register"
+      description="Create an account to start streaming."
+      footerText="Already have an account?"
+      footerLink="/login"
+      footerLinkText="Login"
     >
       <form className="grid gap-4" onSubmit={handleSubmit}>
+        <div className="grid gap-2">
+          <Label htmlFor="username" className="text-muted-foreground">
+            Username
+          </Label>
+          <Input
+            id="username"
+            placeholder="Your username"
+            value={username}
+            onChange={(e) => setUsername(e.target.value)}
+            disabled={isLoading}
+          />
+        </div>
         <div className="grid gap-2">
           <Label htmlFor="email" className="text-muted-foreground">
             Email
@@ -57,7 +70,7 @@ export default function LoginPage() {
         </div>
         <Button type="submit" className="w-full" disabled={isLoading}>
           {isLoading && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
-          Login
+          Create Account
         </Button>
         <GoogleButton />
       </form>
