--- Client/streamverse-app/src/app/auth/route.ts
@@ -1,5 +1,0 @@
-import { redirect } from 'next/navigation';
-
-export async function GET() {
-  redirect('/auth/login');
-}
