+++ Client/streamverse-app/src/app/globals.css
@@ -86,19 +86,19 @@
 
 .dark {
   --background: #111827; /* gray-900 */
-  --foreground: #F9FAFB; /* gray-50 */
-  --card: #1F2937; /* gray-800 */
-  --card-foreground: #F9FAFB; /* gray-50 */
-  --popover: #1F2937; /* gray-800 */
-  --popover-foreground: #F9FAFB; /* gray-50 */
-  --primary: #3B82F6; /* blue-500 */
-  --primary-foreground: #F9FAFB; /* gray-50 */
-  --secondary: #1F2937; /* gray-800 */
-  --secondary-foreground: #F9FAFB; /* gray-50 */
-  --muted: #1F2937; /* gray-800 */
-  --muted-foreground: #9CA3AF; /* gray-400 */
-  --accent: #8B5CF6; /* violet-500 */
-  --accent-foreground: #F9FAFB; /* gray-50 */
+  --foreground: #f9fafb; /* gray-50 */
+  --card: #1f2937; /* gray-800 */
+  --card-foreground: #f9fafb; /* gray-50 */
+  --popover: #1f2937; /* gray-800 */
+  --popover-foreground: #f9fafb; /* gray-50 */
+  --primary: #3b82f6; /* blue-500 */
+  --primary-foreground: #f9fafb; /* gray-50 */
+  --secondary: #1f2937; /* gray-800 */
+  --secondary-foreground: #f9fafb; /* gray-50 */
+  --muted: #1f2937; /* gray-800 */
+  --muted-foreground: #9ca3af; /* gray-400 */
+  --accent: #8b5cf6; /* violet-500 */
+  --accent-foreground: #f9fafb; /* gray-50 */
   --destructive: oklch(0.704 0.191 22.216);
   --border: oklch(1 0 0 / 10%);
   --input: oklch(1 0 0 / 15%);
@@ -132,5 +132,3 @@
     transform: translate(calc(-50% - 0.5rem));
   }
 }
-
-
