+++ Client/streamverse-app/src/app/not-found.tsx
@@ -1,4 +1,3 @@
-
 import Link from "next/link";
 import { Button } from "@/components/ui/button";
