+++ Client/streamverse-app/src/components/stream/HorizontalStreamView.tsx
@@ -16,10 +16,10 @@ export function HorizontalStreamView({ streams }: HorizontalStreamViewProps) {
 
   return (
     <div className="group relative w-full overflow-hidden">
-      <div className="flex animate-scroll space-x-4 group-hover:pause py-4">
+      <div className=" flex animate-scroll space-x-4 group-hover:pause py-4">
         {duplicatedStreams.map((stream, index) => (
-          <div key={`${stream.id}-${index}`}>
-            <StreamCard stream={stream} className="w-[900px]" />
+          <div key={`${stream.id}-${index}`} className="flex-shrink-0">
+            <StreamCard stream={stream} className="w-[400px]" />
           </div>
         ))}
       </div>
